requirements.txt
absl-py==2.1.0
accelerate==1.5.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
alembic==1.15.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
asttokens==3.0.0
attrs==25.3.0
audioread==3.0.1
braceexpand==0.1.7
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.1
cycler==0.12.1
cytoolz==1.0.1
datasets==3.3.2
decorator==5.2.1
dill==0.3.8
Distance==0.1.3
docker-pycreds==0.4.0
docopt==0.6.2
editdistance==0.8.1
einops==0.8.1
executing==2.2.0
fiddle==0.3.0
filelock==3.17.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
future==1.0.0
g2p-en==2.1.0
gitdb==4.0.12
GitPython==3.1.44
graphviz==0.20.3
greenlet==3.1.1
grpcio==1.71.0
huggingface-hub==0.29.3
humanfriendly==10.0
hydra-core==1.3.2
HyperPyYAML==1.2.2
idna==3.10
inflect==7.5.0
intervaltree==3.1.0
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiwer==3.1.0
joblib==1.4.2
julius==0.2.7
kaldi-python-io==1.2.2
kaldiio==2.18.1
kiwisolver==1.4.8
lazy_loader==0.4
Levenshtein==0.27.1
lhotse==1.29.0
libcst==1.7.0
librosa==0.11.0
lightning==2.4.0
lightning-utilities==0.14.0
lilcom==1.8.0
llvmlite==0.44.0
loguru==0.7.3
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapy==1.1.6
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
nemo-toolkit==2.2.0
networkx==3.4.2
nltk==3.9.1
numba==0.61.0
numpy==2.1.3
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.21.0
optuna==4.2.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
peft==0.14.0
pillow==11.1.0
plac==1.4.3
platformdirs==4.3.6
pooch==1.8.2
primePy==1.3
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==3.20.3
psutil==7.0.0
pure_eval==0.2.3
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==19.0.1
pybind11==2.13.6
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pyloudnorm==0.1.1
pyparsing==3.2.1
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.0.post0
pytorch-metric-learning==2.8.1
pytz==2025.1
PyYAML==6.0.2
RapidFuzz==3.12.2
regex==2024.11.6
requests==2.32.3
resampy==0.4.3
rich==13.9.4
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
sacremoses==0.1.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
semver==3.0.4
sentencepiece==0.2.0
sentry-sdk==2.22.0
setproctitle==1.3.5
setuptools==76.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sortedcontainers==2.4.0
soundfile==0.13.1
sox==1.5.0
soxr==0.5.0.post1
speechbrain==1.0.2
SQLAlchemy==2.0.39
stack-data==0.6.3
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.5.0
text-unidecode==1.3
texterrors==0.5.1
threadpoolctl==3.6.0
tokenizers==0.21.1
toolz==1.0.0
torch==2.6.0
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.6.0
torchmetrics==1.6.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.3
typeguard==4.4.2
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wandb==0.19.8
wcwidth==0.2.13
webdataset==0.2.111
Werkzeug==3.1.3
wget==3.2
win32_setctime==1.2.0
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
